+++ src/extension.ts
@@ -16,8 +16,6 @@ if (DEBUG) {
 const activeDocuments = new Map<vscode.TextDocument, {
     closeActiveSubdocumentWithReason(reason: string): Promise<void>
 }>();
-// Tracks all subdocuments whose documents are still open, for potential reuse
-const subdocumentToDocument = new Map<vscode.TextDocument, vscode.TextDocument>();
 
 export function activate(_context: vscode.ExtensionContext) {
 
@@ -87,43 +85,32 @@ export function activate(_context: vscode.ExtensionContext) {
         //     value: cursorPosition.line - viewPortTopPosition.line
         // });
 
-        // Create subdocument with chosen language/extension. Uses an existing subdocument if available.
-        // Could be made configurable depending on template tag, keybinding, etc.
-        let _subdoc: vscode.TextDocument | undefined;
-
-        // There are only a few subdocuments active per workspace, so iterating them all is fine.
-        // This is a later workaround for openTextDocument not working the same on all platforms.
-        // Without this, subdocument would not be reused on reopen.
-        subdocumentToDocument.forEach((anyDoc, anySubdoc) => {
-            if (anyDoc === doc && anySubdoc.languageId === language) {
-                _subdoc = anySubdoc;
-            }
-        });
-
-        // Only one active subdocument per document allowed for simplicity. But there may be several template languages in a document,
-        // so potentially there are multiple subdocuments ready for reuse per document.
+        // Only one active subdocument per document allowed for simplicity.
         if (activeDocuments.has(doc)) {
             await activeDocuments.get(doc)!.closeActiveSubdocumentWithReason('Subdocument closed. This virtual document can be closed.');
         }
         activeDocuments.set(doc, { async closeActiveSubdocumentWithReason() { } });
 
+        // Create subdocument with chosen language/extension. Always creates a new document for simplicity.
+        // Could be made configurable depending on template tag, keybinding, etc.
+        let _subdoc: vscode.TextDocument | undefined;
+
         if (!_subdoc) {
             if (createUntitled) {
                 // This form is not in typescript definitions but is documented here https://code.visualstudio.com/docs/extensionAPI/vscode-api#workspace.openTextDocument
-                // Unfortunately it always creates a new untitled file
+                // It always creates a new untitled file.
                 _subdoc = await (vscode.workspace.openTextDocument as any)({ language }) as vscode.TextDocument;
             } else {
                 // This works usually nicely, reusing the same subdocument for same source, but may give invalid document on some platforms?
                 const filepath = doc.fileName + '.virtual.' + language; // Needs path too? Don't want to save it...
                 // _subdoc = await vscode.workspace.openTextDocument(vscode.Uri.file(filepath).with({ scheme: 'untitled' })); // Not actually untitled as has a bogus filename
                 // See https://github.com/Microsoft/vscode/issues/723#issuecomment-252411918
-                _subdoc = await vscode.workspace.openTextDocument(vscode.Uri.parse('untitled:/' + filepath)); // Not actually untitled as has a bogus filename
+                _subdoc = await vscode.workspace.openTextDocument(vscode.Uri.parse('untitled:/' + filepath)); // Not actually untitled as has a bogus filename, but helps keep track of tab names
             }
         }
 
         // Make typescript narrow the above happily
         const subdoc: vscode.TextDocument = _subdoc;
-        subdocumentToDocument.set(subdoc, doc);
 
         // Open editor in side by side view
         const subeditor = await vscode.window.showTextDocument(subdoc, editor.viewColumn === vscode.ViewColumn.One ? vscode.ViewColumn.Two : vscode.ViewColumn.One);
@@ -201,7 +188,6 @@ export function activate(_context: vscode.ExtensionContext) {
             at: 'center'
         });
 
-
         /**
          * Handlers
          */
@@ -226,6 +212,7 @@ export function activate(_context: vscode.ExtensionContext) {
         //         disposeSubdocument('Document options changed. This virtual document can be closed.');
         //     }
         // });
+        // It would be nice if saving the subdocument could be interrupted and the original would be saved instead.
 
         const throttledDocumentSync = throttle(async () => {
             try {
@@ -270,74 +257,50 @@ export function activate(_context: vscode.ExtensionContext) {
             changeOrigin = 'dispose';
             changeListener.dispose();
 
-            if (vscode.workspace.textDocuments.indexOf(doc) === -1) {
-                // Source document closed, clean everything
-                documentCloseListener.dispose();
-                activeDocuments.delete(doc);
-                // Other subdocuments will receive the same event and close themselves, close this one up
-                subdocumentCloseListener.dispose();
-                subdocumentToDocument.delete(subdoc);
-                await markSubdocumentAsTainted(reason);
-                return;
-            }
+            documentCloseListener.dispose();
+            subdocumentCloseListener.dispose();
 
-            if (vscode.workspace.textDocuments.indexOf(subdoc) === -1) {
-                // Subdocument closed, others may still be open
-                subdocumentCloseListener.dispose();
-                subdocumentToDocument.delete(subdoc);
-                let noSubdocsLeft = true;
-                subdocumentToDocument.forEach(anyDoc => {
-                    if (anyDoc === doc) {
-                        noSubdocsLeft = false;
-                    }
-                });
-                if (noSubdocsLeft) {
-                    // Last one cleans up the main doc
-                    documentCloseListener.dispose();
-                    activeDocuments.delete(doc);
-                }
-                return;
-            }
+            activeDocuments.delete(doc);
 
-            // Otherwise just taint this subdocument but keep the close listeners and document link so that it can be reused if needed.
-            // NOTE: Reusing the same untitled document this way may keep lots of large closures in memory, as this extension parses the source on each invocation,
-            // but it is just a proof-of-concept after all.
-            await markSubdocumentAsTainted(reason);
+            // Experimental closing via action, moves focus so may pipe quick keypresses to wrong doc unfortunately
+            await closeSubdocument(vscode.window.activeTextEditor);
 
-            // Experimental closing via action, moves focus so may pipe keypresses to wrong doc
-            // await closeSubdocument(vscode.window.activeTextEditor);
+            // Alternatively just mark the document as tainted.
+            // await markSubdocumentAsTainted(reason);
+
+            // TODO: perhaps there could be a status bar widget of some sort that would allow easy closing of subdocuments.
         }
 
-        async function markSubdocumentAsTainted(_reason: string) {
-            if (vscode.workspace.textDocuments.indexOf(subdoc) >= 0) {
-                let newSubeditor = await vscode.window.showTextDocument(subdoc, subeditor.viewColumn, /* preserveFocus */ true);
-                try {
-                    let ok = await newSubeditor.edit(builder => {
-                        const totalRange = subdoc.validateRange(new vscode.Range(0, 0, 100000, 100000));
-                        // builder.replace(totalRange, reason || 'This virtual editor can be closed.');
-                        // Return to empty document so that the document won't be marked as dirty and can be closed quickly.
-                        builder.replace(totalRange, '');
-                    });
-                    if (!ok) {
-                        throw new Error('Dispose edit could not succeed');
-                    }
-                } catch (err) {
-                    if (DEBUG) {
-                        console.log('DISPOSE ERR %s', err && err.stack || err);
-                    }
-                }
-            }
-        }
+        // async function markSubdocumentAsTainted(_reason: string) {
+        //     if (vscode.workspace.textDocuments.indexOf(subdoc) >= 0) {
+        //         try {
+        //             let newSubeditor = await vscode.window.showTextDocument(subdoc, subeditor.viewColumn, /* preserveFocus */ true);
+        //             let ok = await newSubeditor.edit(builder => {
+        //                 const totalRange = subdoc.validateRange(new vscode.Range(0, 0, 100000, 100000));
+        //                 // builder.replace(totalRange, reason || 'This virtual editor can be closed.');
+        //                 // Return to empty document so that the document won't be marked as dirty and can be closed quickly.
+        //                 builder.replace(totalRange, '');
+        //             });
+        //             if (!ok) {
+        //                 throw new Error('Dispose edit could not succeed');
+        //             }
+        //         } catch (err) {
+        //             if (DEBUG) {
+        //                 console.log('DISPOSE ERR %s', err && err.stack || err);
+        //             }
+        //         }
+        //     }
+        // }
 
         // No proper way to close the subeditor, other than deprecated editor.hide() method or using actions to close editors.
         // This tries to show the subeditor and then close it via command. Needs to suppress save dialog by clearing the doc,
         // and focusing back to the original editor.
-        // Internally just editorService.closeEditor(position, input);
-        async function closeSubdocument(returnToEditor: vscode.TextEditor) {
+        // Internally just editorService.closeEditor(position, input).
+        async function closeSubdocument(returnToEditor?: vscode.TextEditor) {
             if (vscode.workspace.textDocuments.indexOf(subdoc) >= 0) {
-                // Move focus temporarily to subdocument
-                let newSubeditor = await vscode.window.showTextDocument(subdoc, subeditor.viewColumn, /* preserveFocus */ false);
                 try {
+                    // TODO: subdocument may be visible in multiple editors, this closes just one of them.
+                    let newSubeditor = await vscode.window.showTextDocument(subdoc, subeditor.viewColumn, /* preserveFocus */ true);
                     let ok = await newSubeditor.edit(builder => {
                         const totalRange = subdoc.validateRange(new vscode.Range(0, 0, 100000, 100000));
                         builder.replace(totalRange, ''); // Return to empty document to prevent save dialog
@@ -345,9 +308,13 @@ export function activate(_context: vscode.ExtensionContext) {
                     if (!ok) {
                         throw new Error('Dispose edit could not succeed')
                     }
+                    // Move focus temporarily to subdocument. Do this here to minimize time for the focus to be in wrong doc as the user is typing.
+                    await vscode.window.showTextDocument(subdoc, subeditor.viewColumn, /* preserveFocus */ false);
                     await vscode.commands.executeCommand('workbench.action.closeActiveEditor');
                     // Move focus back to where it was
-                    await vscode.window.showTextDocument(returnToEditor.document, returnToEditor.viewColumn, /* preserveFocus */ false);
+                    if (returnToEditor) {
+                        await vscode.window.showTextDocument(returnToEditor.document, returnToEditor.viewColumn, /* preserveFocus */ false);
+                    }
                 } catch (err) {
                     if (DEBUG) {
                         console.log('DISPOSE ERR %s', err && err.stack || err);
@@ -361,8 +328,9 @@ export function activate(_context: vscode.ExtensionContext) {
     }
 }
 
-// export async function deactivate(_context: vscode.ExtensionContext) {
-//     for (let fn of activeDocuments.values()) {
-//         await fn.dispose('Extension deactivated. This virtual document can be closed.');
-//     }
-// }
+// Cleanup on exit
+export async function deactivate(_context: vscode.ExtensionContext) {
+    for (let handle of activeDocuments.values()) {
+        await handle.closeActiveSubdocumentWithReason('Extension deactivated. This virtual document can be closed.');
+    }
+}